rename socket and unix writer classes
