+++ spec/wavefront-sdk/writers/proxy_spec.rb
@@ -11,36 +11,34 @@ require_relative '../../spec_helper'
 require_relative '../../../lib/wavefront-sdk/write'
 require_relative '../../../spec/support/mocket'
 
-UNIX_SOCK = '/tmp/testsock'
+WS_CREDS = { proxy: 'wavefront-proxy' }.freeze
 
-WU_CREDS = { socket: UNIX_SOCK }.freeze
-
-# Test UNIX Datagram socket writing
+# The Proxy class writes to a proxy TCP socket
 #
-class WavefrontWriterUnixTest < MiniTest::Test
+class WavefrontWriterSocketTest < MiniTest::Test
   attr_reader :wf, :wf_noop
 
   def setup
-    @wf = Wavefront::Write.new(WU_CREDS, writer: :unix)
-    @wf_noop = Wavefront::Write.new(WU_CREDS, writer: :unix, noop: true)
+    @wf = Wavefront::Write.new(WS_CREDS, writer: :proxy)
+    @wf_noop = Wavefront::Write.new(WS_CREDS, writer: :proxy, noop: true)
   end
 
   def test_writer_class
-    assert_instance_of(Wavefront::Writer::Unix, wf.writer)
+    assert_instance_of(Wavefront::Writer::Proxy, wf.writer)
   end
 
   def test_write_openclose
     mocket = Mocket.new
-    Spy.on(UNIXSocket, :new).and_return(mocket)
-    mocket_spy = Spy.on(mocket, :write)
+    Spy.on(TCPSocket, :new).and_return(mocket)
+    mocket_spy = Spy.on(mocket, :puts)
     wf.write(POINT)
     assert mocket_spy.has_been_called?
   end
 
   def test_write_noop
     mocket = Mocket.new
-    Spy.on(UNIXSocket, :new).and_return(mocket)
-    mocket_spy = Spy.on(mocket, :write)
+    Spy.on(TCPSocket, :new).and_return(mocket)
+    mocket_spy = Spy.on(mocket, :puts)
     wf_noop.open
     wf_noop.write(POINT, false)
     refute mocket_spy.has_been_called?
@@ -48,16 +46,16 @@ class WavefrontWriterUnixTest < MiniTest::Test
 
   def test_write_noop_openclose
     mocket = Mocket.new
-    Spy.on(UNIXSocket, :new).and_return(mocket)
-    mocket_spy = Spy.on(mocket, :write)
+    Spy.on(TCPSocket, :new).and_return(mocket)
+    mocket_spy = Spy.on(mocket, :puts)
     wf_noop.write(POINT)
     refute mocket_spy.has_been_called?
   end
 
   def test_write
     mocket = Mocket.new
-    Spy.on(UNIXSocket, :new).and_return(mocket)
-    mocket_spy = Spy.on(mocket, :write)
+    Spy.on(TCPSocket, :new).and_return(mocket)
+    mocket_spy = Spy.on(mocket, :puts)
     wf.open
     wf.write(POINT, false)
     assert mocket_spy.has_been_called?
@@ -65,8 +63,8 @@ class WavefrontWriterUnixTest < MiniTest::Test
 
   def test_write_array
     mocket = Mocket.new
-    Spy.on(UNIXSocket, :new).and_return(mocket)
-    mocket_spy = Spy.on(mocket, :write)
+    Spy.on(TCPSocket, :new).and_return(mocket)
+    mocket_spy = Spy.on(mocket, :puts)
     wf.open
     wf.write(POINT_A, false)
     assert mocket_spy.has_been_called?
@@ -74,32 +72,32 @@ class WavefrontWriterUnixTest < MiniTest::Test
 
   def test_noop_send_point
     mocket = Mocket.new
-    Spy.on(UNIXSocket, :new).and_return(mocket)
-    mocket_spy = Spy.on(mocket, :write)
+    Spy.on(TCPSocket, :new).and_return(mocket)
+    mocket_spy = Spy.on(mocket, :puts)
     wf_noop.open
     wf_noop.send_point(POINT_L)
     refute mocket_spy.has_been_called?
   end
 
   def test_open
-    tcp_spy = Spy.on(UNIXSocket, :new)
+    tcp_spy = Spy.on(TCPSocket, :new)
     wf.open
     assert tcp_spy.has_been_called?
-    assert_equal([UNIX_SOCK], tcp_spy.calls.first.args)
+    assert_equal(['wavefront-proxy', 2878], tcp_spy.calls.first.args)
   end
 
   def test_noop_open
-    tcp_spy = Spy.on(UNIXSocket, :new)
+    tcp_spy = Spy.on(TCPSocket, :new)
     log_spy = Spy.on(wf_noop.logger, :log)
     wf_noop.open
     refute tcp_spy.has_been_called?
-    assert_equal(['No-op requested. Not opening socket connection.'],
+    assert_equal(['No-op requested. Not opening connection to proxy.'],
                  log_spy.calls.last.args)
     assert_equal(1, log_spy.calls.size)
   end
 
   def test_noop_close
-    tcp_spy = Spy.on(UNIXSocket, :new)
+    tcp_spy = Spy.on(TCPSocket, :new)
     log_spy = Spy.on(wf_noop.logger, :log)
     wf_noop.close
     refute tcp_spy.has_been_called?
@@ -107,25 +105,25 @@ class WavefrontWriterUnixTest < MiniTest::Test
   end
 
   def test_validate_credentials
-    assert(Wavefront::Write.new(WU_CREDS, writer: :unix))
+    assert(Wavefront::Write.new(WS_CREDS, writer: :proxy))
 
     assert_instance_of(Wavefront::Write,
-                       Wavefront::Write.new(WU_CREDS, writer: :unix))
+                       Wavefront::Write.new(WS_CREDS, writer: :proxy))
 
     assert_raises(Wavefront::Exception::CredentialError) do
-      Wavefront::Write.new({}, writer: :unix)
+      Wavefront::Write.new({}, writer: :proxy)
     end
 
     assert_raises(Wavefront::Exception::CredentialError) do
-      Wavefront::Write.new({ endpoint: 'wavefront.com' }, writer: :unix)
+      Wavefront::Write.new({ endpoint: 'wavefront.com' }, writer: :proxy)
     end
 
     assert_raises(Wavefront::Exception::CredentialError) do
-      Wavefront::Write.new({ token: 'abcdef' }, writer: :unix)
+      Wavefront::Write.new({ token: 'abcdef' }, writer: :proxy)
     end
 
     assert_raises(Wavefront::Exception::CredentialError) do
-      Wavefront::Write.new({ proxy: nil }, writer: :unix)
+      Wavefront::Write.new({ proxy: nil }, writer: :proxy)
     end
   end
 end
